pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptclaw"
version = "2.1.0"
description = "An orchestrated multi-agent prompt scaffold with artifact-based handoffs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
coherence-pg = ["psycopg2-binary>=2.9", "pgvector>=0.2", "redis>=5.0"]
coherence-full = ["psycopg2-binary>=2.9", "pgvector>=0.2", "redis>=5.0", "pyyaml>=6.0"]

[project.scripts]
promptclaw = "promptclaw.cli:main"

[tool.setuptools.packages.find]
include = ["promptclaw*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
